[feature] complete CountdownScreen

+++ feature/countdown/src/main/java/ir/kaaveh/countdown/model/CountdownState.kt
@@ -3,5 +3,5 @@ package ir.kaaveh.countdown.model
 data class CountdownState(
     val counterState: CounterState = CounterState.INITIAL,
     val workingState: WorkingState = WorkingState.REST,
-    val remainTime: Int = WORKING_DURATION,
+    val remainTime: Long = WORKING_DURATION,
 )
