+++ library/ext-functions/src/main/java/ir/kaaveh/ext_functions/Time.kt
@@ -1,9 +1,9 @@
 package ir.kaaveh.ext_functions
 
-val Int.minutes: Int
+val Long.minutes: Long
     get() = this / 60
 
-val Int.seconds: Int
+val Long.seconds: Long
     get() = this % 60
 
 fun String.toTowDigitFormat(): String =
